backend/requirements.txt
asgiref==3.8.1
boto3==1.37.32
botocore==1.37.32
Django==5.1.4
django-cors-headers==4.6.0
django-tenants==3.7.0
djangorestframework==3.15.2
jmespath==1.0.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.4.0